fix: The GetCanaries endpoint returns with the same error for all namespaces
We already add an extra error to the response if Flagger is not
installed on a cluster, so we can ignore all errors from that cluster
and we will only return with the "not installed" error.

If Flagger is installed, but something is wrong, it will still return
the same error for all namespaces. Right now, I can't think of a case
that would yield this result.

Drawback: if the error is not related to Flagger and Canary objects, it
will still suppress the error messages and show only the "not installed"
error. It shouldn't be an issue, because if Flagger is not installed,
but something else blew up, it's not an endpoint related issue, but it's
most likely a cluster setup issue.

Fixes #947